modify exception handler and add some bugs :bug:...
