+++ src/main/java/com/nvd/footballmanager/controller/BaseController.java
@@ -1,14 +1,10 @@
 package com.nvd.footballmanager.controller;
 
 import com.nvd.footballmanager.dto.CustomApiResponse;
-import com.nvd.footballmanager.exceptions.AccessDeniedException;
-import com.nvd.footballmanager.exceptions.BadRequestException;
 import com.nvd.footballmanager.filters.BaseFilter;
 import com.nvd.footballmanager.service.BaseService;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.MappedSuperclass;
 import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -60,22 +56,6 @@ public abstract class BaseController<E, DTO, FT extends BaseFilter, ID extends U
     public ResponseEntity<CustomApiResponse> delete(@PathVariable("id") ID id) {
         baseService.deleteById(id);
         return ResponseEntity.ok(CustomApiResponse.noContent());
-    }
-
-    @ExceptionHandler({AccessDeniedException.class})
-    public ResponseEntity<CustomApiResponse> handleAccessDenied(AccessDeniedException ex) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(CustomApiResponse.forbidden(ex.getMessage()));
-    }
-
-
-    @ExceptionHandler({EntityNotFoundException.class})
-    public ResponseEntity<CustomApiResponse> handleEntityNotFound(EntityNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(CustomApiResponse.notFound(ex.getMessage()));
-    }
-
-    @ExceptionHandler({BadRequestException.class})
-    public ResponseEntity<CustomApiResponse> handleBadRequest(BadRequestException ex) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(CustomApiResponse.badRequest(ex.getMessage()));
     }
 
 }
