+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=FootballManager
 #server.port=8000
-#spring.datasource.url=jdbc:mysql://localhost:3306/football_manager
-spring.datasource.url=jdbc:mysql://localhost:3306/football_manager_seed
+spring.datasource.url=jdbc:mysql://localhost:3306/football_manager
+#spring.datasource.url=jdbc:mysql://localhost:3306/football_manager_seed
 spring.datasource.username=root
 spring.datasource.password=dinhhtvq123
 #spring.datasource.driver-class-name=com.mysql.jdbc.Driver
